+++ .eslintrc.json
@@ -12,6 +12,8 @@
   "extends": ["airbnb-base"],
   "rules": {
     // "linebreak-style": ["error", "linux"],
+    // "import/no-default-export": "error",
+    "import/prefer-default-export": "off",
     "max-len": ["error", { "code": 500 }],
     "max-classes-per-file": "off",
     "no-shadow": "off",
