+++ src/index.html
@@ -26,6 +26,9 @@
       <button class="clear">Clear all completed</button>
     </section>
 
-    <script src="./index.js" type="application/json"></script>
+    <script src="./index.js" type="module"></script>
+    <script src="./modules/toDoList.js" type="module"></script>
+    <script src="./modules/localStorage.js" type="module"></script>
+    <script src="./modules/userInterface.js" type="module"></script>
   </body>
 </html>
