+++ src/modules/toDoList.js
@@ -1,0 +1,9 @@
+class ToDoList {
+  constructor(index, desc, comp) {
+    this.index = index;
+    this.desc = desc;
+    this.comp = comp;
+  }
+}
+
+export default ToDoList;
